backend/almacen/requirements.txt
asgiref==3.9.2
defusedxml==0.7.1
diff-match-patch==20241021
Django==4.0.4
django-cors-headers==3.11.0
django-filter==21.1
django-import-export==2.7.1
djangorestframework==3.13.1
et_xmlfile==2.0.0
odfpy==1.4.1
openpyxl==3.1.5
pillow==11.3.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
xlrd==2.0.2
xlwt==1.3.0
gunicorn==21.2.0
whitenoise==6.5.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
